requirements.txt
Flask==3.1.2
mysql-connector-python==9.5.0
Werkzeug==3.1.3